+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.1] - 2025-05-25
+
+- Added ghost-import command to convert Ghost blog exports to willow.camp markdown format
+- Support for converting Ghost exports with proper frontmatter
+- Automatically detect and use the best content format available (markdown, HTML or plaintext)
+
 ## [0.1.0] - 2025-05-25
 
 - Initial release